Updates following Mike's comments
